refactor: move dates to lib folder
