+++ src/app/admin/api-keys/[name]/page.tsx
@@ -7,7 +7,7 @@ import DateInput from '@/components/UI/DateInput'
 import TextInput from '@/components/UI/TextInput'
 import DisplayAllPermissions from '@/components/Permission/DisplayAllPermissions'
 import Slider from '@/components/UI/Slider'
-import { displayDate } from '@/dates/displayDate'
+import { displayDate } from '@/lib/dates/displayDate'
 import { destroyApiKeyAction } from '@/actions/api-keys/destroy'
 
 type PropTypes = {
