+++ src/app/events/[order]/[name]/page.tsx
@@ -4,7 +4,7 @@ import CreateOrUpdateEventForm from '@/app/events/CreateOrUpdateEventForm'
 import { readEvent } from '@/actions/events/read'
 import CmsImage from '@/app/_components/Cms/CmsImage/CmsImage'
 import CmsParagraph from '@/app/_components/Cms/CmsParagraph/CmsParagraph'
-import { displayDate } from '@/dates/displayDate'
+import { displayDate } from '@/lib/dates/displayDate'
 import Form from '@/components/Form/Form'
 import EventTag from '@/components/Event/EventTag'
 import { destroyEvent } from '@/actions/events/destroy'
